# components/backend/src/main/resources/restaurant-api-doc.yaml
openapi: 3.0.3
info:
  title: Swagger Restaurant
  description: |-
    This is a sample Restaruant
  version: 1.0.1
...
servers:
  - url: http://localhost:8084
...
paths:
...
  /booking:
    get:
      tags:
        - booking
      summary: Query bookings
      description: 
      operationId: queryBookings
      security:
      - sessionId: []
      parameters:
        - name: table-number
          in: header
          schema: { type: integer }
          required: false
        - name: customer-name
          in: header
          schema: { type: string }
          required: false
        - name: from
          in: header
          schema:
            type: string
            format: date-time
          required: false
        - name: to
          in: header
          schema:
            type: string
            format: date-time
          required: false
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
                      
        '405':
          description: Invalid input
    post:
      tags:
        - booking
      summary: Book table
      operationId: bookTable
      security:
      - sessionId: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewBooking'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'

        '405':
          description: Invalid input
  /booking/{id}:
    parameters:
      - name: id
        in: path
        schema: { type: string }
        required: true
    get:
      tags:
        - booking
      summary: Get booking
      operationId: getBooking
      security:
      - sessionId: []
      responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Booking'

          '405':
            description: Invalid input
    patch:
      tags:
        - booking
      summary: Update booking status
      operationId: bookingStatusUpdate
      security:
      - sessionId: []
      parameters:
        - name: status
          in: header
          schema: 
            type: string
            enum:
              - pending
              - accepted
          required: true
      responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Booking'
          '405':
            description: Invalid input
  /booking/table/avaible:
    get:
      tags:
        - booking
      summary: Query avaible tables
      description: 
      operationId: queryTablesForBooking
      security:
      - sessionId: []
      parameters:
        - name: from
          in: header
          schema:
            type: string
            format: date-time
          required: true
        - name: to
          in: header
          schema:
            type: string
            format: date-time
          required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Table'
                      
        '405':
          description: Invalid input
...
components:
  schemas:
    Table:
      required:
        - number
        - capacity
      type: object
      properties:
        number:
          type: integer
          format: int32
          example: 10
        capacity:
          type: integer
          format: int32
          example: 10
        status:
          type: string
          enum: [booked, free]
    Booking:
      required:
        - id
        - table
        - customer
        - startingDate
        - endingDate
        - status
        - peopleCount
      type: object
      properties:
        id:
          type: string
          example: ObjectId('6544cd596955fe0a1c04fba9')
        table:
          $ref: '#/components/schemas/Table'
        customer:
          $ref: '#/components/schemas/User'
        startingDate:
          type: string
          format: date-time
        endingDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, accepted]
        peopleCount:
          type: integer
    NewBooking:
      required:
        - table-number
        - customer-id
        - startingDate
        - endingDate
        - peopleCount
      type: object
      properties:
        table-number:
          type: integer
          example: 3
        customer-id:
          type: string
          example: 6544cd596955fe0a1c04fba9
        startingDate:
          type: string
          format: date-time
        endingDate:
          type: string
          format: date-time
        peopleCount:
          type: integer
...
    User:
      required:
        - name
        - email
        - mobil
        - address
      type: object
      properties:
        id:
          type: string
          example: ObjectId('6544cd596955fe0a1c04fba9')
        name:
          type: string
          example: Gipsz Jakap
        email:
          type: string
          example: gipsz@jakab.com
        mobil:
          type: string
          example: +36709834234
        address:
          type: string
          example: Budapest Lakatos utca 6.
...
  securitySchemes:
    sessionId:
      type: apiKey
      in: header
      name: sessionId